+++ src/output/worldStateManager/WorldStateManager.cpp
@@ -268,6 +268,158 @@ bool WorldStateManager::IsSkyChase1Enabled()
     return true;
 }
 
+void WorldStateManager::DrawDisableDoorIndicator(const NJS_POINT3 basePoint, const float angle)
+{
+    NJS_POINT3COL point3Col;
+    NJS_POINT3 point[] = {
+        {basePoint.x, basePoint.y + CROSS_SIZE_MIN, basePoint.z},
+        {basePoint.x + CROSS_SIZE_DIFF, basePoint.y + CROSS_SIZE_MAX, basePoint.z},
+        {basePoint.x + CROSS_SIZE_MAX, basePoint.y + CROSS_SIZE_DIFF, basePoint.z},
+        {basePoint.x + CROSS_SIZE_MIN, basePoint.y, basePoint.z},
+        {basePoint.x + CROSS_SIZE_MAX, basePoint.y - CROSS_SIZE_DIFF, basePoint.z},
+        {basePoint.x + CROSS_SIZE_DIFF, basePoint.y - CROSS_SIZE_MAX, basePoint.z},
+        {basePoint.x, basePoint.y - CROSS_SIZE_MIN, basePoint.z},
+        {basePoint.x - CROSS_SIZE_DIFF, basePoint.y - CROSS_SIZE_MAX, basePoint.z},
+        {basePoint.x - CROSS_SIZE_MAX, basePoint.y - CROSS_SIZE_DIFF, basePoint.z},
+        {basePoint.x - CROSS_SIZE_MIN, basePoint.y, basePoint.z},
+        {basePoint.x - CROSS_SIZE_MAX, basePoint.y + CROSS_SIZE_DIFF, basePoint.z},
+        {basePoint.x - CROSS_SIZE_DIFF, basePoint.y + CROSS_SIZE_MAX, basePoint.z},
+    };
+    // Rotation angle in radians
+    const float theta = angle * (3.14 / 180.0f);
+    const float cosTheta = cos(theta);
+    const float sinTheta = sin(theta);
+
+    // Rotate each point around the Y axis
+    for (auto& p : point)
+    {
+        float x = p.x - basePoint.x;
+        float z = p.z - basePoint.z;
+
+        p.x = x * cosTheta + z * sinTheta + basePoint.x;
+        p.z = -x * sinTheta + z * cosTheta + basePoint.z;
+    }
+
+    point3Col.p = point;
+    point3Col.tex = nullptr;
+    point3Col.col = this->_wrongDoorColor;
+    late_DrawPolygon3D(&point3Col, 12, NJD_TRANSPARENT, LATE_LIG);
+}
+
+void WorldStateManager::DrawCorrectDoorIndicator(const NJS_POINT3 basePoint, const float angle)
+{
+    NJS_POINT3COL point3Col;
+    NJS_POINT3 point[] = {
+        {basePoint.x, basePoint.y + ARROW_SIZE_MAX, basePoint.z},
+        {basePoint.x + ARROW_SIZE_MIN, basePoint.y + ARROW_SIZE_MAX - ARROW_SIZE_MIN, basePoint.z},
+        {basePoint.x + ARROW_SIZE_MIN / 3, basePoint.y + ARROW_SIZE_MAX - ARROW_SIZE_MIN, basePoint.z},
+        {basePoint.x + ARROW_SIZE_MIN / 3, basePoint.y - ARROW_SIZE_MAX, basePoint.z},
+        {basePoint.x - ARROW_SIZE_MIN / 3, basePoint.y - ARROW_SIZE_MAX, basePoint.z},
+        {basePoint.x - ARROW_SIZE_MIN / 3, basePoint.y + ARROW_SIZE_MAX - ARROW_SIZE_MIN, basePoint.z},
+        {basePoint.x - ARROW_SIZE_MIN, basePoint.y + ARROW_SIZE_MAX - ARROW_SIZE_MIN, basePoint.z},
+    };
+    // Rotation angle in radians
+    const float theta = angle * (3.14 / 180.0f);
+    const float cosTheta = cos(theta);
+    const float sinTheta = sin(theta);
+
+    // Rotate each point around the Y axis
+    for (auto& p : point)
+    {
+        float x = p.x - basePoint.x;
+        float z = p.z - basePoint.z;
+
+        p.x = x * cosTheta + z * sinTheta + basePoint.x;
+        p.z = -x * sinTheta + z * cosTheta + basePoint.z;
+    }
+
+    const EntityData1* player = EntityData1Ptrs[0];
+    const double dz = basePoint.z - player->Position.z;
+    const double dy = basePoint.y - player->Position.y;
+    const double dx = basePoint.x - player->Position.x;
+    const double distance = sqrt(dx * dx + dy * dy + dz * dz);
+
+    float extraPercentage;
+    if (distance <= MIN_DRAW_DOOR_ARROW_DISTANCE)
+        extraPercentage = 0;
+    else if (distance >= MAX_DRAW_DOOR_ARROW_DISTANCE)
+        extraPercentage = 1;
+    else
+        extraPercentage = (distance - MIN_DRAW_DOOR_ARROW_DISTANCE) / (MAX_DRAW_DOOR_ARROW_DISTANCE -
+            MIN_DRAW_DOOR_ARROW_DISTANCE);
+
+    uint8_t newAlpha = static_cast<uint8_t>(0xAA - 0xAA * (1 - extraPercentage));
+    _arrowColor.argb.a = newAlpha;
+
+    NJS_COLOR arrowColor[7] = {
+        _arrowColor,
+        _arrowColor,
+        _arrowColor,
+        _arrowColor,
+        _arrowColor,
+        _arrowColor,
+        _arrowColor,
+    };
+
+    point3Col.p = point;
+    point3Col.tex = nullptr;
+    point3Col.col = arrowColor;
+    late_DrawPolygon3D(&point3Col, 7, NJD_TRANSPARENT, LATE_LIG);
+}
+
+void WorldStateManager::DrawOtherDoorIndicator(const NJS_POINT3 basePoint, const float angle)
+{
+    NJS_POINT3COL point3Col;
+    NJS_POINT3 point[] = {
+        {basePoint.x + ARROW_SIZE_MAX, basePoint.y, basePoint.z},
+        {basePoint.x + ARROW_SIZE_MAX - ARROW_SIZE_MIN, basePoint.y + ARROW_SIZE_MIN, basePoint.z},
+        {basePoint.x + ARROW_SIZE_MAX - ARROW_SIZE_MIN, basePoint.y + ARROW_SIZE_MIN / 3, basePoint.z},
+        {basePoint.x - ARROW_SIZE_MAX, basePoint.y + ARROW_SIZE_MIN / 3, basePoint.z},
+        {basePoint.x - ARROW_SIZE_MAX, basePoint.y - ARROW_SIZE_MIN / 3, basePoint.z},
+        {basePoint.x + ARROW_SIZE_MAX - ARROW_SIZE_MIN, basePoint.y - ARROW_SIZE_MIN / 3, basePoint.z},
+        {basePoint.x + ARROW_SIZE_MAX - ARROW_SIZE_MIN, basePoint.y - ARROW_SIZE_MIN, basePoint.z},
+    };
+    // Rotation angle in radians
+    const float theta = angle * (3.14 / 180.0f);
+    const float cosTheta = cos(theta);
+    const float sinTheta = sin(theta);
+
+    // Rotate each point around the Y axis
+    for (auto& p : point)
+    {
+        float x = p.x - basePoint.x;
+        float z = p.z - basePoint.z;
+
+        p.x = x * cosTheta + z * sinTheta + basePoint.x;
+        p.z = -x * sinTheta + z * cosTheta + basePoint.z;
+    }
+    point3Col.p = point;
+    point3Col.tex = nullptr;
+    point3Col.col = this->_otherDoorColor;
+    late_DrawPolygon3D(&point3Col, 7, NJD_TRANSPARENT, LATE_LIG);
+}
+
+void WorldStateManager::ShowLevelEntranceArrows()
+{
+    if (!this->options.entranceRandomizer)
+        return;
+    if (!this->_showEntranceIndicators)
+        return;
+    for (LevelArrow levelArrow : _levelArrows)
+    {
+        if (CurrentStageAndAct != levelArrow.levelAndAct)
+            continue;
+
+        if (!worldStateManagerPtr->levelEntrances.canEnter(levelArrow.entrance, CurrentCharacter))
+            DrawDisableDoorIndicator(levelArrow.postion, levelArrow.angle);
+        else if (levelArrow.isForCharacter(static_cast<Characters>(CurrentCharacter)))
+            DrawCorrectDoorIndicator(levelArrow.postion, levelArrow.angle);
+        else
+            DrawOtherDoorIndicator(levelArrow.postion, levelArrow.angle);
+
+    }
+}
+
 void WorldStateManager::OnFrame()
 {
     if (DemoPlaying > 0)
@@ -278,6 +430,7 @@ void WorldStateManager::OnFrame()
     if (IsSkyChase1Enabled())
         EventFlagArray[33] = 1;
 
+    this->ShowLevelEntranceArrows();
 
     if (this->eggCarrierTransformationCutscene)
     {
@@ -572,6 +725,11 @@ void WorldStateManager::SetChaoStatsMultiplier(const int chaoStatsMultiplier)
     }
 }
 
+void WorldStateManager::SetShowEntranceIndicators(const bool showEntranceIndicators)
+{
+    this->_showEntranceIndicators = showEntranceIndicators;
+}
+
 typedef struct
 {
     int x;
@@ -764,7 +922,7 @@ FunctionHook<void> onCountSetItemsMaybe(0x0046BD20, []()-> void
     if (worldStateManagerPtr->options.skyChaseChecks)
     {
         //Sky Chase
-        if(worldStateManagerPtr->IsSkyChase1Enabled())
+        if (worldStateManagerPtr->IsSkyChase1Enabled())
         {
             AddSetToLevel(WARP_SKY_CHASE_1, LevelAndActIDs_MysticRuins1, Characters_Sonic);
             AddSetToLevel(WARP_SKY_CHASE_1, LevelAndActIDs_MysticRuins1, Characters_Tails);